Also define invokeDslMethod() and getDslProperty() on DslScript's meta class, use DslDelegate as class for MissingMethodExceptions.
invokeDslMethod() and getDslProperty() allows to redefine methodMissing
and propertyMissing() on the DslScripts' meta class, which can e.g. be
useful when providing some APIs to the DSL script.

Using DslDelegate for MissingMethodExceptions makes more clear what
happened. Just telling the user that the method was missing is much more
confusing. Consider using addSuppressed().